Restructure pages for user flow based on user type.
