+++ pages/listings/index.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+import { Listings } from '../../components';
+
+class NewListingPage extends React.Component {
+  render() {
+    return (
+      <div>
+        <Listings />
+      </div>
+    );
+  }
+}
+
+export default NewListingPage;
+
